oneshot_guessle/requirements/top_level_deps.txt
argon2-cffi==21.3.0
async-timeout==4.0.2
black==23.1.0
crispy-bootstrap5==0.7
cryptography==40.0.2
django-allauth==0.52.0
django-coverage-plugin==3.0.0
django_debug_toolbar==3.8.1
django-environ==0.9.0
django-extensions==3.2.1
django-htmx==1.14.0
django-model-utils==4.3.1
django-redis==5.2.0
django-social-share==2.3.0
django-stubs==1.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
flake8-isort==6.0.0
fontawesomefree==6.4.0
hiredis==3.2.1
importlib-metadata==6.5.0
ipdb==0.13.11
pillow==11.2.1
pipdeptree==2.26.1
pre-commit==3.0.4
psycopg2-binary==2.9.10
pylint-django==2.5.3
pytest-django==4.5.2
pytest-sugar==0.9.6
python-slugify==8.0.0
pytz==2022.7.1
sphinx-autobuild==2021.3.14
watchdog==3.0.0
Werkzeug==3.1.3
whitenoise==6.3.0